Connect to remote host asynchronously
Connecting to the remote host in init/1 blocks the gen_server process.
If the connect fails, the gen_server must handle the error case or crash.
Instead, add an extra state and timeout immediately in init/1.

A badly behaving client could cause a large number of connection attempts
to be made. The gen_server could keep a count in its state or put the
state in an ets table.