Preparation for checking session transmitted bytes
Modify session record and state to maintain transmitted bytes in both
directions. If the number of bytes is not kept, lost packets will disrupt
the session.

When the client is sending data, the byte count should be equal to our
count plus the size of the new data. Otherwise, the packet is assumed
to be a duplicate and will be dropped.

When the client is requesting pending data, the byte count should be
equal to ours. If it is smaller, we should re-send our last packet.